+++ src/autostore/types/__init__.py
@@ -1,5 +1,19 @@
 """Types."""
 
-from .sqlalchemy import FloatArrayTypeDecorator, PathTypeDecorator
+from .fields import Role
+from .sqlalchemy import (
+    FloatArrayTypeDecorator,
+    ModelT,
+    PathTypeDecorator,
+    RowID,
+    RowIDs,
+)
 
-__all__ = ["FloatArrayTypeDecorator", "PathTypeDecorator"]
+__all__ = [
+    "Role",
+    "FloatArrayTypeDecorator",
+    "ModelT",
+    "PathTypeDecorator",
+    "RowID",
+    "RowIDs",
+]